+++ app/src/main/java/com/undistract/ProfilesPicker.kt
@@ -36,6 +36,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.compose.ui.window.Dialog
 import com.undistract.data.models.Profile
+import com.undistract.data.repositories.ProfileManagerRepository
 
 
 @Composable
