+++ app/src/main/java/com/undistract/data/repositories/ProfileManagerRepository.kt
@@ -1,8 +1,9 @@
-package com.undistract
+package com.undistract.data.repositories
 
 import android.content.Context
 import android.content.SharedPreferences
 import android.util.Log
+import com.undistract.AppInfo
 import com.undistract.data.models.Profile
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -40,7 +41,7 @@ class ProfileManagerRepository(private val context: Context) {
                 val profilesList = mutableListOf<Profile>()
 
                 for (i in 0 until jsonArray.length()) {
-                    profilesList.add(Profile.fromJson(jsonArray.getJSONObject(i)))
+                    profilesList.add(Profile.Companion.fromJson(jsonArray.getJSONObject(i)))
                 }
 
                 _profiles.value = profilesList
