+++ app/src/main/java/com/undistract/services/BlockerService.kt
@@ -1,11 +1,10 @@
-package com.undistract
+package com.undistract.services
 
 import android.app.Service
 import android.content.Intent
 import android.os.IBinder
 import android.util.Log
 import androidx.localbroadcastmanager.content.LocalBroadcastManager
-import com.undistract.services.AppBlockerAccessibilityService
 
 class BlockerService : Service() {
     companion object {
